refactor(workspace): extract injectLiveCwd to shared util, patch WorkspacePicker save sites
- Create src/renderer/src/lib/workspace-utils.ts with shared injectLiveCwd
- Remove duplicate injectLiveCwd definitions from App.tsx and Sidebar.tsx
- Apply injectLiveCwd transform to all 3 layout.save() calls in WorkspacePicker:
  commitNewWorkspace, handleSaveCurrent, handleSwitchWorkspace
